Test independently released CE version and development CE branch
The previous config was testing:
- the last tag of CE branch 4.0.x
- and then the dev branch 4.0

As everything was in the same jobs, there were some bugs.

For example, yarn was not re-installed when modifying composer to load the development CE branch.
If a new dependency was added in the dev CE branch, it was failing due to this installation bug of the CI.

Splitting it in two jobs allows us to:
- test the release version exactly the same way as we document how to create a new project in the documentation
- test the CE branch without relying of previously installed stuff, which can be incomplete/not up to date